Move recovery code model to sub-folder
